chore: rename onchain bria handlers
