+++ src/app/wallets/index.ts
@@ -1,6 +1,6 @@
 export * from "./add-invoice-for-wallet"
-export * from "./add-pending-transaction"
-export * from "./add-settled-transaction"
+export * from "./add-pending-on-chain-transaction"
+export * from "./add-settled-on-chain-transaction"
 export * from "./create-on-chain-address"
 export * from "./get-balance-for-wallet"
 export * from "./get-last-on-chain-address"
